Rework of Middle Eastern sky cultures
Update current Arabic sky culture and add new ones

Arabic Lunar Stations
This is a complete rework of the old Arabic Moon Stations.
The sky culture's importance is its widespread use until the present.

Arabian Peninsula
This is the oral folk star names and asterisms of the Arabian Peninsula.

Arabic
This is the indigenous traditional ancient sky culture of the Arabs. It is completely different and a replacement of the current Arabic sky culture.

Al-Sufi
This is sky culture reflecting the content of Book of the Stars by Al-Sufi written in ~ 964. This part displays the Almagest star layer of the book.

Al-Sufi Arabic
This sky culture displays the indigenous Arabic star names and asterisms layer of Book of the Stars by Al-Sufi. The lines are that of the Arabic asterisms superimposed on the images of the classical Ptolemaic constellations.

Both Al-Sufi and Al-Sufi Arabic have the images of the current Arabic sky culture with corrections and modifications.