Add special permission to distribute Kamilaroi in Stellarium
The permission is stated like this in an email sent to Fabien on
Feb 27th 2021 by Robert Fuller:

"Authorisation
I authorise the Stellarium developers to offer the Kamilaroi/Euahlayi
Skycultures package as a paid-for-extension pack with the CC BY-NC-ND 4.0 license.
Signed,
Michael Eckford/Anderson

Authorisation
I authorise the Stellarium developers to offer the Kamilaroi/Euahlayi
Skycultures package as a paid-for-extension pack with the CC BY-NC-ND 4.0 license.
Signed,
Robert S. Fuller
"